refactor: move assets, ast, auth, files, items, graphql and webhooks types from api to shared
Move those types to have them available for services interfaces in shared
