+++ api/src/controllers/collections.ts
@@ -3,7 +3,7 @@ import { ForbiddenException } from '../exceptions';
 import { respond } from '../middleware/respond';
 import { validateBatch } from '../middleware/validate-batch';
 import { CollectionsService, MetaService } from '../services';
-import { Item } from '../types';
+import { Item } from '@directus/shared/types';
 import asyncHandler from '../utils/async-handler';
 
 const router = Router();
