+++ api/src/services/authorization.ts
@@ -1,6 +1,12 @@
 import { FailedValidationException } from '@directus/shared/exceptions';
 import { AbstractServiceOptions } from '@directus/shared/services';
 import {
+	AST,
+	FieldNode,
+	FunctionFieldNode,
+	Item,
+	NestedCollectionNode,
+	PrimaryKey,
 	Accountability,
 	Aggregate,
 	Filter,
@@ -14,7 +20,6 @@ import { Knex } from 'knex';
 import { cloneDeep, flatten, isArray, isNil, merge, reduce, uniq, uniqWith } from 'lodash';
 import getDatabase from '../database';
 import { ForbiddenException } from '../exceptions';
-import { AST, FieldNode, FunctionFieldNode, Item, NestedCollectionNode, PrimaryKey } from '../types';
 import { stripFunction } from '../utils/strip-function';
 import { ItemsService } from './items';
 import { PayloadService } from './payload';
