+++ api/src/services/files.ts
@@ -1,5 +1,6 @@
 import formatTitle from '@directus/format-title';
 import { AbstractServiceOptions } from '@directus/shared/services';
+import { File, PrimaryKey, MutationOptions, Metadata } from '@directus/shared/types';
 import axios, { AxiosResponse } from 'axios';
 import exifr from 'exifr';
 import { clone, pick } from 'lodash';
@@ -14,7 +15,6 @@ import env from '../env';
 import { ForbiddenException, InvalidPayloadException, ServiceUnavailableException } from '../exceptions';
 import logger from '../logger';
 import storage from '../storage';
-import { File, PrimaryKey, MutationOptions, Metadata } from '../types';
 import { toArray } from '@directus/shared/utils';
 import { ItemsService } from './items';
 import net from 'net';
