+++ api/src/services/flows.ts
@@ -1,7 +1,6 @@
 import { AbstractServiceOptions } from '@directus/shared/services';
-import { FlowRaw } from '@directus/shared/types';
+import { FlowRaw, Item, MutationOptions, PrimaryKey } from '@directus/shared/types';
 import { getFlowManager } from '../flows';
-import { Item, MutationOptions, PrimaryKey } from '../types';
 import { ItemsService } from './items';
 
 export class FlowsService extends ItemsService<FlowRaw> {
