+++ api/src/services/graphql/index.ts
@@ -1,6 +1,15 @@
 import { BaseException } from '@directus/shared/exceptions';
 import { AbstractServiceOptions } from '@directus/shared/services';
-import { Accountability, Action, Aggregate, Filter, Query, SchemaOverview } from '@directus/shared/types';
+import {
+	Accountability,
+	Action,
+	Aggregate,
+	Filter,
+	GraphQLParams,
+	Item,
+	Query,
+	SchemaOverview,
+} from '@directus/shared/types';
 import argon2 from 'argon2';
 import {
 	ArgumentNode,
@@ -52,7 +61,6 @@ import getDatabase from '../../database';
 import env from '../../env';
 import { ForbiddenException, GraphQLValidationException, InvalidPayloadException } from '../../exceptions';
 import { getExtensionManager } from '../../extensions';
-import { GraphQLParams, Item } from '../../types';
 import { generateHash } from '../../utils/generate-hash';
 import { getGraphQLType } from '../../utils/get-graphql-type';
 import { reduceSchema } from '../../utils/reduce-schema';
