+++ api/src/services/items.ts
@@ -1,5 +1,15 @@
 import { AbstractService, AbstractServiceOptions } from '@directus/shared/services';
-import { Accountability, Action, PermissionsAction, Query, SchemaOverview } from '@directus/shared/types';
+import {
+	Accountability,
+	Action,
+	PermissionsAction,
+	Query,
+	SchemaOverview,
+	ActionEventParams,
+	Item as AnyItem,
+	MutationOptions,
+	PrimaryKey,
+} from '@directus/shared/types';
 import Keyv from 'keyv';
 import { Knex } from 'knex';
 import { assign, clone, cloneDeep, omit, pick, without } from 'lodash';
@@ -10,7 +20,6 @@ import emitter from '../emitter';
 import env from '../env';
 import { ForbiddenException, InvalidPayloadException } from '../exceptions';
 import { translateDatabaseError } from '../exceptions/database/translate';
-import { ActionEventParams, Item as AnyItem, MutationOptions, PrimaryKey } from '../types';
 import getASTFromQuery from '../utils/get-ast-from-query';
 import { validateKeys } from '../utils/validate-keys';
 import { AuthorizationService } from './authorization';
