+++ api/src/services/payload.ts
@@ -1,5 +1,14 @@
 import { AbstractServiceOptions } from '@directus/shared/services';
-import { Accountability, Query, SchemaOverview } from '@directus/shared/types';
+import {
+	Accountability,
+	Query,
+	SchemaOverview,
+	ActionEventParams,
+	Alterations,
+	Item,
+	MutationOptions,
+	PrimaryKey,
+} from '@directus/shared/types';
 import { format, parseISO, isValid } from 'date-fns';
 import { parseJSON, toArray } from '@directus/shared/utils';
 import { unflatten } from 'flat';
@@ -11,7 +20,6 @@ import { parse as wktToGeoJSON } from 'wellknown';
 import getDatabase from '../database';
 import { getHelpers, Helpers } from '../database/helpers';
 import { ForbiddenException, InvalidPayloadException } from '../exceptions';
-import { ActionEventParams, Alterations, Item, MutationOptions, PrimaryKey } from '../types';
 import { generateHash } from '../utils/generate-hash';
 import { ItemsService } from './items';
