+++ api/src/services/relations.ts
@@ -2,7 +2,7 @@ import { Knex } from 'knex';
 import { systemRelationRows } from '../database/system-data/relations';
 import { ForbiddenException, InvalidPayloadException } from '../exceptions';
 import { AbstractServiceOptions } from '@directus/shared/services';
-import { SchemaOverview, Relation, RelationMeta, Accountability, Query } from '@directus/shared/types';
+import { ActionEventParams, MutationOptions, SchemaOverview, Relation, RelationMeta, Accountability, Query } from '@directus/shared/types';
 import { toArray } from '@directus/shared/utils';
 import { ItemsService, QueryOptions } from './items';
 import { PermissionsService } from './permissions';
@@ -12,7 +12,6 @@ import getDatabase, { getSchemaInspector } from '../database';
 import { getDefaultIndexName } from '../utils/get-default-index-name';
 import { getCache, clearSystemCache } from '../cache';
 import Keyv from 'keyv';
-import { ActionEventParams, MutationOptions } from '../types';
 import { getHelpers, Helpers } from '../database/helpers';
 import emitter from '../emitter';
 import { getSchema } from '../utils/get-schema';
