+++ api/src/services/tfa.ts
@@ -4,7 +4,7 @@ import { authenticator } from 'otplib';
 import getDatabase from '../database';
 import { InvalidPayloadException } from '../exceptions';
 import { ItemsService } from './items';
-import { PrimaryKey } from '../types';
+import { PrimaryKey } from '@directus/shared/types';
 
 export class TFAService {
 	knex: Knex;
