+++ api/src/services/users.ts
@@ -6,9 +6,8 @@ import env from '../env';
 import { FailedValidationException } from '@directus/shared/exceptions';
 import { ForbiddenException, InvalidPayloadException, UnprocessableEntityException } from '../exceptions';
 import { RecordNotUniqueException } from '../exceptions/database/record-not-unique';
-import { Item, PrimaryKey, MutationOptions } from '../types';
 import { AbstractServiceOptions } from '@directus/shared/services';
-import { Query, SchemaOverview, Accountability } from '@directus/shared/types';
+import { Query, SchemaOverview, Accountability, Item, PrimaryKey, MutationOptions } from '@directus/shared/types';
 import isUrlAllowed from '../utils/is-url-allowed';
 import { toArray } from '@directus/shared/utils';
 import { Url } from '../utils/url';
