+++ api/src/services/webhooks.ts
@@ -1,5 +1,5 @@
 import { AbstractServiceOptions } from '@directus/shared/services';
-import { Item, PrimaryKey, Webhook, MutationOptions } from '../types';
+import { Item, PrimaryKey, Webhook, MutationOptions } from '@directus/shared/types';
 import { ItemsService } from './items';
 import { getMessenger, Messenger } from '../messenger';
