+++ api/src/types/index.ts
@@ -1,13 +1,7 @@
-export * from './assets';
-export * from './ast';
 export * from './auth';
 export * from './collection';
 export * from './events';
-export * from './files';
-export * from './graphql';
-export * from './items';
 export * from './meta';
 export * from './migration';
 export * from './revision';
 export * from './snapshot';
-export * from './webhooks';
