+++ api/src/utils/get-ast-from-query.ts
@@ -4,8 +4,16 @@
 
 import { Knex } from 'knex';
 import { cloneDeep, mapKeys, omitBy, uniq, isEmpty } from 'lodash';
-import { AST, FieldNode, FunctionFieldNode, NestedCollectionNode } from '../types';
-import { Query, PermissionsAction, Accountability, SchemaOverview } from '@directus/shared/types';
+import {
+	AST,
+	FieldNode,
+	FunctionFieldNode,
+	NestedCollectionNode,
+	Query,
+	PermissionsAction,
+	Accountability,
+	SchemaOverview,
+} from '@directus/shared/types';
 import { getRelationType } from '../utils/get-relation-type';
 import { REGEX_BETWEEN_PARENS } from '@directus/shared/constants';
