+++ packages/shared/package.json
@@ -73,6 +73,7 @@
 		"express": "4.18.2",
 		"fs-extra": "10.1.0",
 		"geojson": "0.5.0",
+		"graphql": "16.6.0",
 		"joi": "17.7.0",
 		"knex": "2.3.0",
 		"knex-schema-inspector": "3.0.0",
@@ -90,6 +91,7 @@
 		"@types/geojson": "7946.0.10",
 		"@types/lodash": "4.14.189",
 		"@types/pino": "7.0.4",
+		"@types/sharp": "0.31.0",
 		"@types/tmp": "0.2.3",
 		"concurrently": "7.5.0",
 		"tmp": "0.2.1",
