+++ packages/shared/src/types/assets.ts
