+++ packages/shared/src/types/auth.ts
@@ -1,0 +1,6 @@
+export type LoginResult = {
+	accessToken: any;
+	refreshToken: any;
+	expires: any;
+	id?: any;
+};
