+++ packages/shared/src/types/index.ts
@@ -1,14 +1,19 @@
-export * from './activity';
 export * from './accountability';
+export * from './activity';
+export * from './assets';
+export * from './ast';
+export * from './auth';
 export * from './collection';
 export * from './displays';
 export * from './endpoints';
 export * from './events';
 export * from './extensions';
 export * from './fields';
+export * from './files';
 export * from './filter';
 export * from './flows';
 export * from './geometry';
+export * from './graphql';
 export * from './hooks';
 export * from './interfaces';
 export * from './items';
@@ -27,3 +32,4 @@ export * from './settings';
 export * from './shares';
 export * from './users';
 export * from './vue';
+export * from './webhooks';
